+++ src/models/builders/EntrySiteRelationsBuilder.php
@@ -6,6 +6,7 @@ use davidhirtz\yii2\cms\models\Asset;
 use davidhirtz\yii2\cms\models\Entry;
 use davidhirtz\yii2\cms\models\SectionEntry;
 use davidhirtz\yii2\cms\modules\ModuleTrait;
+use davidhirtz\yii2\media\models\File;
 use davidhirtz\yii2\skeleton\helpers\ArrayHelper;
 use Yii;
 use yii\base\BaseObject;
@@ -15,9 +16,17 @@ class EntrySiteRelationsBuilder extends BaseObject
     use ModuleTrait;
 
     public Entry $entry;
+
+    /** @var Asset[] */
     public array $assets = [];
+
+    /** @var Entry[] */
     public array $entries = [];
 
+    /** @var File[] */
+    public array $files = [];
+
+    protected array $fileIds = [];
     protected array $relatedEntryIds = [];
     protected array $sectionIdsWithAssets = [];
     protected array $sectionIdsWithEntries = [];
@@ -25,6 +34,8 @@ class EntrySiteRelationsBuilder extends BaseObject
     public function init(): void
     {
         ArrayHelper::index($this->entries, 'id');
+        ArrayHelper::index($this->files, 'id');
+
         $this->entries[$this->entry->id] = $this->entry;
 
         $this->loadRelations();
@@ -40,6 +51,8 @@ class EntrySiteRelationsBuilder extends BaseObject
         $this->loadEntries();
 
         $this->loadAssets();
+        $this->loadFiles();
+
         $this->populateAssetRelations();
     }
 
@@ -77,7 +90,7 @@ class EntrySiteRelationsBuilder extends BaseObject
             return;
         }
 
-        Yii::debug('Loading entries for sections ...');
+        Yii::debug('Loading section entry relations ...');
 
         /** @var SectionEntry[] $sectionEntries */
         $sectionEntries = SectionEntry::find()
@@ -109,16 +122,13 @@ class EntrySiteRelationsBuilder extends BaseObject
         if ($entryIds) {
             Yii::debug('Loading related entries ...');
 
-            $this->entries = [
-                ...$this->entries,
-                ...Entry::find()
-                    ->selectSiteAttributes()
-                    ->replaceI18nAttributes()
-                    ->whereStatus()
-                    ->andWhere(['id' => $entryIds])
-                    ->indexBy('id')
-                    ->all()
-            ];
+            $this->entries += Entry::find()
+                ->selectSiteAttributes()
+                ->replaceI18nAttributes()
+                ->whereStatus()
+                ->andWhere(['id' => $entryIds])
+                ->indexBy('id')
+                ->all();
         }
     }
 
@@ -158,14 +168,39 @@ class EntrySiteRelationsBuilder extends BaseObject
         $this->assets = Asset::find()
             ->selectSiteAttributes()
             ->replaceI18nAttributes()
-            ->withFiles()
             ->whereStatus()
             ->andWhere(count($condition) > 1 ? ['or', ...$condition] : $condition[0])
             ->all();
+
+        foreach ($this->assets as $asset) {
+            $this->fileIds[] = $asset->file_id;
+        }
+    }
+
+    protected function loadFiles(): void
+    {
+        $fileIds = array_unique($this->fileIds);
+        $fileIds = array_diff($fileIds, array_keys($this->files));
+
+        Yii::debug('Loading related files ...');
+
+        if ($fileIds) {
+            $this->files += File::find()
+                ->selectSiteAttributes()
+                ->replaceI18nAttributes()
+                ->where(['id' => $fileIds])
+                ->withFolder()
+                ->indexBy('id')
+                ->all();
+        }
     }
 
     protected function populateAssetRelations(): void
     {
+        foreach ($this->assets as $asset) {
+            $asset->populateFileRelation($this->files[$asset->file_id] ?? null);
+        }
+
         foreach ($this->entries as $entry) {
             $entry->populateAssetRelations($this->assets);
         }