+++ src/views/site/view.php
@@ -12,5 +12,5 @@ use davidhirtz\yii2\cms\widgets\MetaTags;
 use davidhirtz\yii2\cms\widgets\Sections;
 use davidhirtz\yii2\skeleton\web\View;
 
-MetaTags::register(['model' => $entry]);
-echo Sections::widget(['entry' => $entry,]);
+MetaTags::widget(['model' => $entry]);
+echo Sections::widget(['entry' => $entry]);